Moving exports to addon-test-support
